// frontend/src/sections/asset-catalogue/reasons/logReasons.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type AssetLogReasonsVariables = {
  storeId: string;
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    assetLogStatus?: {
    equalTo?: "NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE" | null;
    equalAny?: Array<"NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE"> | null;
    notEqualTo?: "NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE" | null;
    notEqualAll?: Array<"NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE"> | null;
  } | null;
    reason?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
  } | null;
  sort?: Array<{
    key: "status";
    desc?: boolean | null;
  }> | null;
};

export type AssetLogReasonsResult = {
  assetLogReasons: ({
  __typename: "AssetLogReasonConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  assetLogStatus: "NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE";
  reason: string;
  commentsRequired: boolean;
}>;
});
};

export const AssetLogReasons = {
  query: "query assetLogReasons($storeId: String!, $filter: AssetLogReasonFilterInput, $sort: [AssetLogReasonSortInput!]) {\n  assetLogReasons(storeId: $storeId, filter: $filter, sort: $sort) {\n    ... on AssetLogReasonConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        assetLogStatus\n        reason\n        commentsRequired\n      }\n    }\n  }\n}",
} as TypedDocument<AssetLogReasonsResult, AssetLogReasonsVariables>;

export type InsertAssetLogReasonVariables = {
  input: {
    id: string;
    assetLogStatus: "NOT_IN_USE" | "FUNCTIONING" | "FUNCTIONING_BUT_NEEDS_ATTENTION" | "NOT_FUNCTIONING" | "DECOMMISSIONED" | "UNSERVICEABLE";
    reason: string;
    commentsRequired: boolean;
  };
};

export type InsertAssetLogReasonResult = {
  centralServer: {
  logReason: {
  insertAssetLogReason: ({
  __typename: "AssetLogReasonNode";
} & {
  id: string;
});
};
};
};

export const InsertAssetLogReason = {
  query: "mutation insertAssetLogReason($input: InsertAssetLogReasonInput!) {\n  centralServer {\n    logReason {\n      insertAssetLogReason(input: $input) {\n        __typename\n        ... on AssetLogReasonNode {\n          id\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<InsertAssetLogReasonResult, InsertAssetLogReasonVariables>;

export type DeleteLogReasonVariables = {
  id: string;
};

export type DeleteLogReasonResult = {
  centralServer: {
  logReason: {
  deleteLogReason: ({
  __typename: "DeleteResponse";
} & {
  id: string;
});
};
};
};

export const DeleteLogReason = {
  query: "mutation deleteLogReason($id: String!) {\n  centralServer {\n    logReason {\n      deleteLogReason(reasonId: $id) {\n        __typename\n        ... on DeleteResponse {\n          id\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<DeleteLogReasonResult, DeleteLogReasonVariables>;